Add configuration by user config file
Configuration of fetchsep codes is moved from the config.py module
that the user was expected to edit, to configuration files in the
INI format used by python configparser.  Config files are handled
in a precedence order:

 1. fetchsep.cfg in the current working directory
 2. .fetchsep in the user's home directory
 3. fetchsep.cfg in the directory of this package

By default, the user should not have to specify any config file.
One only needs create a config file to specify non-default values.

This fixes the problem in the pip deployment where the user
generally does not have access to the config.py previously used.

config.py is still the method of supplying config variables to
the opsep/idsep codes.  It has been rewritten to parse the
text config files and export into the package namespace.
Downstream use should be relatively seamless.