Addressed robustness issues
When pulling GOES integral fluxes from iSWA, an empty csv file will be
created for dates that don't exist yet. This can results in empty files
in the data directory. Because the files are present, the code (read_datasets)
will not write over them when the data becomes available. When downloading
from iSWA, added that wget returns a filename. If the data is present, the
filename is 'data'. If the data is not present, there is no filename and
this will cause an error. Since the download is in a try block, this effectively
skips creating an empty datafile.

Due to the fact that empty files were being read into idsep and opsep,
a number of crashes occurred because the code was not robust enough to
check for empty lists or arrays. Some modifications were made to
perform such checks.

A similar crash occurred when inputting the same start and end time into
opsep via the fetchsep_prepare_obs workflow. This caused crashes that
were fixed by checking for empty arrays.