Finalize prepsep, improve idsep, code cleanup
The previous move_output.py was changed to prepsep and updated to have the following capabilities:
-Print out the existing observation windows in the target directory and the approved SEP list
-Run opsep with a startdate equal to the last observation window of the observations in the target directory
-If an SEP event is encountered, the files will not be moved unless the event is in the approved SEP list
-If there is no SEP, files are moved automatically

The above modifications to prepsep should enable integration into the SPHINX validation framework for semi-automated running.

Code throughout the repository was cleaned up with better syntax. Some subroutines were moved to
different files in utils, particularly if there were redundant copies in both opsep and idsep.

A slightly different workflow was implemented in idsep in an attempt to get a
better background flux estimation. Added an increased number of iterations to identify the
background and enhanced fluxes. This was done differently that before by using the SEP periods
to separate enhancements from background. In the very last iteration, the final SEP fluxes
and background fluxes are found by excluding the time periods found to be SEP events from
the original input fluxes and then calculating the background (i.e. putting back in all the points
that had previously been excluded as high points in an effort to get a more realistic background).
The increased number of iterations corrected problems with background identification at the start
of the analyzed time period, particularly when enhanced fluxes occurred right at the beginning
of the data set.

The increased number of required iterations to separate background and SEP causes increased
computational time. Some efforts were begun to optimize the subroutines in derive_background_idsep.py,
however they were not yet completed and have been commented out at the bottom of the file.

For GOES_RT data, added the rerequest subroutine from Luke Stegeman which makes multiple requests for
a url in the case of a hang. iSWA was having problems with hanging, so this solved the issue. This subroutine
is only applied to the GOES_RT data at the moment.