prepsep: Create and populate approved_SEP file if doesn't exist
Add the feature to check if approved_SEP.csv exists in the target directory where
opsep files will be moved. Files containing an SEP event can only be moved to the target
directory if they are in the approved list. If the approved list does not exist, the new
default behavior is to check all of the observations already in the target directory and populate
the approved_SEP.csv file with this SEP events upon initialization.

Made the TwoPeaks flag as an input to prepsep.