Optimize sliding window algorithm in idsep
In define_background_idsep.py, some subroutines were removed because they
were not used in the origina background identification subroutines and they
were definitely not needed for the optimized background identification subroutines.

Algorithms were written to optimize the background identification. This is still
in a testing state, so the old algorithms are commented out for testing and comparison.
The new ndays_average() optimized subroutine takes advantage of pandas dataframes.
The new algorithm uses df.std() instead of np.std(), which give slightly different results
when checked on the same input. The new algorithm gives results that are consistent
with the previous approach, but the final numbers for the background and threshold
values are not exactly the same. Testing continues to determine whether the two algorithms
are consistent enough. SEP Times are almost exactly the same across the two approaches,
however some differences are present in some cases. Testing on the reasons behind
this continues.

Subroutines to write out dataframes or txt files from the two ndays_average()
approaches were added for comparison of results.

idsep: apply_sliding_window algorithm changed to take an iteration
number to append to plot names.