idsep can write fluxes to csv file
write_fluxes subroutine is added to tools.py to write out fluxes to
csv file with datetime in first column and fluxes in remaining columns.
The energy bins are the headers for each column.

A workflow was created in idsep to download/read in observational data,
address any bad values through interpolation or setting to None, then
writing out the data to csv file. This gives the user an easy-to-read
file with the exact data used for analysis in idsep.