Choice of original or optimized subroutines in idsep
Added --Optimized flag. If the flag is not set, then the original
routines for background identification and background/SEP separation
are used. If the --Optimized flag is set, then the newly developed
optimized versions of those subroutines will be used.

This was implemented for testing to ensure that the new methods are consistent
with the original methods.