Finalize Optimized routine in idsep
The optimized version of background identification and background/SEP
separation has been tested and compared to the previous inefficient
version of the code and verified. Some bugs were found in the previous method,
but they were largely inconsequential to final results. The upside is that
the optimized routines are verified to do exactly as intended.

Removing the old, inefficient method and keeping only the new method
using dataframes. Removing the Optimize argument.