Add padding to start of data for robustness
The background solution in idsep takes some time to "settle". Testing showed that
starting the data ~4 months prior to the time period of interest delivered a robust
background solution for averaging windows of 5 up to 27 days. Added in a
"padding" of 27*4 days or 4*sliding_win (max choice) prior to the desired
start time to ensure the background solution has settled to the robust values.

Made it the default behavior to write fluxes to file when running idsep.