Added data manager to check for complete data files
Previously, if a file was present on the user's computer, FetchSEP assumes that the
user has the data and does not download the file again. This can lead to incomplete
files if the same data file continued to be updated after it was downloaded, i.e.
yearly files where each new day is added; daily files when the user downloaded the data
before the day was over.

A data/fetchsep_data_manager.csv file keeps track of every file downloaded by
fetchsep with a boolean indicating whether the file was complete. New file
manager and completeness subroutines in read_datasets.py now check
for completeness and will download a file again if it is marked as
incomplete in the fetchsep_data_manager.csv file, ensuring the user
has the most up-to-date data.

Implemented for GOES and GOES_RT. Still need to implement for
GOES_R, EPHIN, and STEREO.