Delete incomplete files and redownload
If a file is in the data/ directory, but found to be incomplete, first delete
the existing file and download again. The deletion step was added to avoid
the OS from saving a copy with an appended filename. A quick test of
wget with the overwrite option (-O) on a Mac did not stop the behavior, so
chose to delete existing file prior to download.

Added a few comments to understand why STEREO data takes so long to load.