Fix major bug in GOES_RT data reading
GOES real time data downloaded from CCMC's iSWA server can have gaps in time.
read_in_goes_RT() didn't properly account for this, resulting in fluxes
shifted with respect to time, giving incorrect timing information. This
problem was fixed taking advantage of dataframes. The gaps in time are
still present in the fluxes as no effort was yet taken to add in the
missing timestamps, however the fluxes with time are now correct.