Implement Primary/Secondary GOES time series
read_datasets.py and idsep have been updated with the capability to read in
all primary or secondary (as specified by user) GOES spacecraft
and string them together into a single time series. This method
references the file reference/GOES_Primary_Secondary_Status.csv
and will currently work between 1986 to 2024-09-06.

The implementation is still imperfect and will be updated
with features to fill gaps when present. There are cases where
proton data exists, but the orientation files are missing. Or only
EPS or EPEAD data exists but not HEPAD, or vice versa. These
cases still need to be handled more gracefully.

GOES-05, -06, and -07 are added and extend back to 1984. GOES
proton data is sparse prior to this date.