Improve robustness of background identification
Some GOES data shows very non-Gaussian flux distributions that cause
errors in identifing the mean background and sigma. The algorithm
applied in finding background values assumes a normal-like distribution.
Checking was added in define_background_idsep.py to determine whether
the data being evaluated as the background has a normal-enough
distribution to be meaningful. A subroutine called
check_bg_for_normality applies scipy.stats.normaltest to
calculate a p value indicating the likelihood that the
distribution was sampled from a normal distribution. A cut was
applied on the p value to exclude time windows that contain
very non-Gaussian distributions from being used to calculate
background mean and sigma. If a bad distribution is identified, the
last good mean and sigma are used instead. This method makes the
background estimation much more robust for GOES experiments.
The p value cut may need to be tweaked/changed for different experiments.
In this update, all p values < 1e-4 are excluded.