Improve robustness of background identification
Some GOES data shows very non-Gaussian flux distributions that cause
errors in identifing the mean background and sigma. The algorithm
applied in finding background values assumes a normal-like distribution.
Checking was added in define_background_idsep.py to determine whether
the data being evaluated as the background has a normal-enough
distribution to be meaningful.

The previous attempt to solve this problem was not so effective.
Instead, the background fluxes are put into a histogram and the
kurtosis is calculated to give an indicator if there are any
outlier bins containing a lot of entries. A cut is placed on the
kurtosis value to exclude time periods when the background
flux distribution is not appropriate for estimating a mean and sigma.

The cut on kurtosis has been made an adjustable parameter added in
config.py as kurtosis_cut. For GOES data with the binning implemented
in define_background_idsep.py, a cut on kurtosis of 50 - 70 works well.
The numerical values may change with data set, flux units, and levels.
The kurtosis with time is output in csv and pkl/kurtosis_it*. This
will allow the user to choose an appropriate value if necessary.