Prepare to incorporate idsep background for SEP onset ID in opsep
Adjust filenames in idsep to indicate the final background solution
with the tag _FINAL as they had been previously labeled by iteration.

Add input booleans throughout opsep and batch run files to indicate
that the user would like to use the mean background and threshold
calculated in idsep as an applied threshold in opsep, with the
intention of identifying the first particle rise above background.

Make some figures in opsep larger and adjust font size.