Produce json file output
OpSEP works from end-to-end for GOES (untested with
other experiments) to produce json files in the correct
format. The test period was 2012-05-17 and there may be
a bug that prevents running other times. More testing
is needed, but OpSEP works from input to output for
at least some cases at this point.