Begin to implement opsep use of idsep bg and thresholds
Goal is to use the mean background and thresholds calculated by
idsep for background subtraction and SEP event definitions.
Starting the process, but not complete.

Fixed bug in do_interpolation that was not treating null
fluxes properly since null values were changed to nan
instead of None.

Changing many of opsep inputs to keywords for easier use
from the command line or as an imported module.