Implement enhancement above background; idsep integration
OpSEP is now able to read idsep output csv files that contain
the idsep mean background and sigma values with time.

--IDSEPEnhancement and --OPSEPEnhancement have been added as
flags to indicate the user wants to find the SEP enhancement
times and associated fluences above background.

--IDSEPSubtractBG specifies to subtract the idsep mean background
with time from the fluxes.

--OPSEPSubtractBG specifies to subtract the mean background with time
calculated from a background time period specified by the user with
--BGStartDate and --BGEndDate.

If any of the above options are chosen, the fluxes deemed to be
background fluxes, flux <= mean+n*sigma, are set to 0.
Fluxes deemed to be SEP fluxes, flux > mean+n*sigma, are left as-is
or subtracted by the mean background is the SubtractBG flags are set.

The number of sigma used for background and SEP separation by OpSEP is
specified in fetchsep.cfg in opsep_nsigma.

To identify the enhancements above background, a variable in fetchsep.cfg,
opsep_min_threshold, is used to set a very low threshold that is
necessarily below the lowest enhancement value. It can be set by
choosing a value below the lowest background intensity in the data set.
Can be set to zero, but will not show up well in plots.

Still need to update batch workflows with new variables.