Tidy ACE, SOHO, IMP8; Make event definition checking robust
There were some bugs with reading and plotting ACE SIS, SOHO EPAM electrons,
and IMP-8 CPME in opsep. These were fixed.

Some of the above datasets do not contain energy channels in the 10 - 100 MeV
range. This caused opsep to fail because it expected to be able to apply
>10 MeV, 10 pfu and >100 MeV, 1 pfu event definitions. OpSEP now gracefully
indicates that those energy ranges aren't present in the data and removes
those event definitions. The user must input an event definition based off
of one of the energy channels present in the data in order to calculate
SEP information.