Module for detecting new onsets
Independent module for automatically detecting onsets. Uses the fluence
spectra in a running 30 minute period and fits to a Band function
(double powerlaw with a smooth transition). From the fit, the code
uses the change in the fit parameters to determine if a new onset has
occurred. Idea is that there is some change in the fit (hardening in
the high energy spectral slope, change in break energy, etc) then an
onset might have occurred.

Notable Blocks:
From_fetchsep: idea is to put all transition pieces from fetchsep here
so that the onset code itself is generic.

function_two (needs a better name tbh): takes the input flux file and
takes the 30 minute running average fluence spectrum, fits it, and
attempts to find new onsets.

fitting_routine: controls the input into the fit functions
(fit_band_function), in theory new fitting functions could be added
(Ellison-Ramaty, powerlaw with exponential decay, etc) so they would
be added here.

All other functions are here to help run the other functions above -
some are even in FetchSEP but I wanted this to not need dependencies
since I wasn't sure where this was going to live in the workflow.