Able to generate SEP list end-to-end using idsep bg in opsep
fetchsep_prepare_obs and all batch run related files have been updated
to allow the end-to-end use of idsep to identify SEP enhancements and
quantify the mean background and sigma. Then to run opsep for each
time period, taking advantage of the idsep mean background and sigma to
remove the background periods, leaving only SEP enhancments. Background
subtraction may be done using the idsep mean background.

This is a first draft of the functionality needed to create the
CLEAR Benchmark dataset with the most basic infrastructure in place.