Address missing data in GOES orientation files
Some GOES-13, -14, and -15 orientation files are missing the last day
of the file. It seems that some months that have 31 days were processed
to only 30 days. Previously, FetchSEP had a list of bad files to
skip. Here the code was fixed to use the orientation of the
previous day to all the code to read all of the data, including
the last day of the month even if the orientation file was incomplete.

The orientation file for 20130501 to 20130531 is missing data and only
contains some html tags. A communication with Juan Rodriguez clarified
that "GOES-13 was always upright starting in 2010." The orientation file
for April 2013 stated that the B detector was the westward-facing detector,
so it is now hard coded that the B detector is the westward-facing detector
for May 2013 so that the correct data could be selected.

Some fixes have been made to the labels, legends, and titles in OpSEP
figures to ensure that all user-selected information is reflected in
the figures and so that titles don't go off of the plot, etc.