Adjust filenames, fix bug in read_in_calgoes
Changed filename of event list compiled when opsep run in batch mode.
Made sure file contained all uniquely identifying strings like the
plots and directory names.

There was a bug in read_in_calgoes that resulted in a dates array
that was one element shorter than the fluxes array likely causing
a shift by one timestep.

Added code to exclude the last day of GOES-16 differential data
on 2025-04-07 because the file appears to have a problem.