Add flags; output pickle SEP list
Added flag in fetchsep_prepare_obs to allow the user to specify
the name of a batch file to be created with a unique name or
to use as input if starting at the BATCH start point. This allows
the user to curate a batch file and use it to run OpSEP on
desired time periods.

Added flag in fetchsep_prepare_obs to specify the reference
energy bin to determine which sep_times file should be used
to create the batch time periods. Previously the variable was
stored in the config file, but is more useful as an input.

Created a new output pickle file for each time period run
by OpSEP and for the full event list run by OpSEP when using
batch mode.