Improve event identification above bg in opsep
To identify SEP events above background in OpSEP, modified
the algorithm used in IDSEP to identify enhancements above
background. To do a more precise identification of single
SEPs, changed some of the parameters in the algorithm
to do a better job differentiating between enhancements
due to SEPs and lots of 3sigma points above background.
This was done in tools.py > identify_sep_above_background_one.
Changes were also made to account for NaN values if interpolation
is not applied.

The original algorithm used in IDSEP is now in
tools.py > identify_sep_above_background(). This algorithm may
need to be modified in a similar approach to the new
subroutine above, but it seems to be working well, even without
interpolation, so leaving it for now.

Add bug fixes throughout to make robust to NaN values when not
using interpolation.

Added logic to identify bad Weibull fits when attempting to find
onset peaks. Will exclude fits that return the default parameters
or parameters that are at the limits of their ranges.