Create download-only workflow; user specified data path
Created a new module that only downloads and plots data between the
user-specified time frame. Made the datapath a command line input
so that the user can download the data to different locations
using a script. FetchSEP will default to the datapath in the
config file if no path is provided.

Download data using bin/download without using any of FetchSEP's
other features.

Specified fonts in plots so that dates angled at 45 degrees will fit
completely on the plot.

Fixed a bug in idsep that was encountered when the data at the
beginning of the file was not suitable for calculating the mean
background.