Working Test Version
Now properly outputs a new batch list to be fed into OPSEP that
includes the initial list but with your manually selected end times
added in (with the next start time being the previous end).

Needs a bit more work to improve efficiency - crashed in my testing of
a large input file of events from 1987 - 1996 (crashed in 1991). Right
now it has the entire flux array saved for each event rather than just
the chunks between its start/end. Need to make that change, but this
is nearly there (i think)