Add flare, cme associations from SRAG list
Added a modified and cleaned version of Steve Johnson's SRAG
SEP event list with all of the associated flare, cme, radio
and other contextual information for SEP events in SC 23 - 25
with some events in SC 22.

Flare and CME trigger blocks are added to output json files
when relevant.

Currently, the code automatically looks for associations in the
SRAG list. Will be made optional in a future version.

Made some modifications to the output csv files containing the
SEP event summary information to make it more readable and
generically ingestible.