Consistent use of config vars; update shell scripts
There were places that config variables were overridden by local
global variables. Make every module always refer to the config
variables and, if needed, change the global config. Right now,
that option only exists with datapath, but can be expanded.

Shell scripts updated to include more detailed comments and in the
process of making access points so the user doesn't have to run the
whole script if not needed.