Main to create GOES primary list; Update CLEAR scripts
Add a function in bin to allow users to call the module
that compiles all GOES lists into a single SEP event list
drawing from the primary spacecraft.

Update Mac and Linux scripts with documentation, time stamps,
and a break point that allows the user to skip the step to
calculate the mean backgrounds with idsep, if it has already
been done previously.