Finalize CLEAR deploy scripts
Update scripts to run time consuming idsep mean background
calculation (12+ hour run time) in one loop and perform the
SEP event analysis and list generation (2+ hour run time)
in another loop. Add the ability to use an
optional argument "LISTS" on the command line to skip the
mean background calculation, if the user has already done it
previously, and go right to reprocessing the SEP events.

Add timestamps throughout the scripts to allow the user to keep
track of progress.

When the script is completed, copy the default fetchsep.cfg
file to the home directory to avoid overwriting files in
the CLEAR/ directory when running fetchsep on the fly.