Add TwoPeaks capability; revise CLEAR batch lists
Add back in the feature to extend an SEP event if there is an initial
increase above threshold, a temporary dip below, and then an increase again.
The user may specify the --TwoPeaks flag for the code to look for a
second increase above threshold and extend the SEP start and end times to
include both increases. Has no effect on event definition if a second
increase isn't present.

Updated CLEAR batch lists with TwoPeaks flag as needed and revised a
couple of start and end times.

Slightly modified the onset peak quality check criteria to exclude a
fit only if ALL fit parameters are set to the default values.
2012-03-07 uses a fit with two parameters at the default value and
one parameter that has been adjusted and the result comes out
appropriately. Trying this set of criteria to check if it gives
good results across the GOES dataset.