Onset peak2 (#29)
* Add criteria to ensure quality of onset peak fit

Exclude fit if Weibull fit:
-parameters are null
- parameters are all the default value
- parameters are the minimum or maximum value in the allowed parameter range
- parameters are within 1e-4 relative difference of minimum/maximum value (close to end of range)
- the identified peak location is at the ends of the fit array
- the minimum of the second derivative is >= 0
- the normalized minimum second derivative is > -0.00013
- the maximum first derivative is less than 1e-3 (flat)
- the normalized first derivative < 0.01 (flat)
- number of Weibull fit points within a factor of 2 of the data points < 65%

If the onset peak is found to be null due to a bad fit, set
equal to the max flux if the max flux occurs with 18 hours
of the SEP event start time

* Add TwoPeaks capability; revise CLEAR batch lists

Add back in the feature to extend an SEP event if there is an initial
increase above threshold, a temporary dip below, and then an increase again.
The user may specify the --TwoPeaks flag for the code to look for a
second increase above threshold and extend the SEP start and end times to
include both increases. Has no effect on event definition if a second
increase isn't present.

Updated CLEAR batch lists with TwoPeaks flag as needed and revised a
couple of start and end times.

Slightly modified the onset peak quality check criteria to exclude a
fit only if ALL fit parameters are set to the default values.
2012-03-07 uses a fit with two parameters at the default value and
one parameter that has been adjusted and the result comes out
appropriately. Trying this set of criteria to check if it gives
good results across the GOES dataset, however found that fits that 
use the default parameters are typically not good and are rejected by
the other criteria outlined above.

* Update CLEAR batch lists with CLEAR in path

* Update CLEAR deploy scripts for Mac, Linux, and Windows to make primary list (The CLEAR List)